# code/webapp/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Waaiburg API",
        "description": "Waaiburg API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/userTypes": {
            "get": {
                "tags": [
                    "userTypes"
                ],
                "summary": "Get list of userTypes",
                "description": "Returns list of userTypes",
                "operationId": "userTypesIndex",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer {token}",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "userTypes": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UserType"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "userTypes"
                ],
                "summary": "Create a new userType",
                "description": "Returns the created userType",
                "operationId": "userTypesStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserType"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User type created succesfully"
                                        },
                                        "userType": {
                                            "$ref": "#/components/schemas/UserType"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/userTypes/{id}": {
            "delete": {
                "tags": [
                    "userTypes"
                ],
                "summary": "Delete an existing userType",
                "description": "Returns the deleted userType",
                "operationId": "userTypesDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of userType to return",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User type deleted succesfully"
                                        },
                                        "userType": {
                                            "$ref": "#/components/schemas/UserType"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "userTypes"
                ],
                "summary": "Update an existing userType",
                "description": "Returns the updated userType",
                "operationId": "userTypesUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of userType to return",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User type updated succesfully"
                                        },
                                        "userType": {
                                            "$ref": "#/components/schemas/UserType"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserType": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Admin"
                    },
                    "description": {
                        "type": "string",
                        "example": "Admin"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-01-01 00:00:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-01-01 00:00:00"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}